♻️ Refactor: 구글 설문 배너 컴포넌트 분리
refactor: 구글 설문 배너 컴포넌트 분리
